requirements.txt
streamlit
sqlalchemy
bcrypt
psycopg2
python-dotenv